TxSchoolCovidTracker_Final/requirements.txt
--find-links wheelhouse
matplotlib
click==8.0.3
distlib==0.3.4
filelock==3.4.2
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
platformdirs==2.5.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
virtualenv==20.13.1
Werkzeug==2.0.3
pandas==1.3.5
